refactor: simplify vscode-lm toggle and remove enable/disable
This commit addresses the review comment on PR #2 to merge the enable and disable methods into a single toggle method.

- Refactored the VSCodeLm class to simplify the toggle logic and remove the enable/disable methods.

- Updated VSCodeHostImpl to expose the toggle method.

- Updated webview.ts interface to include the toggle method.

- Updated useVSCodeLmModels hook to use the toggle method.

- Updated model-section.tsx component to use the toggle method.
